Align button-next examples with icon-button
