+++ packages/wix-ui-core/stories/backoffice/button-next/showcase/affixes.tsx
@@ -2,9 +2,19 @@ import * as React from "react";
 
 import Sound from "wix-ui-icons-common/Sound";
 import CodeShowcase from "wix-storybook-utils/CodeShowcase";
-import { ButtonNext } from "../../../../../src/components/button-next";
-import { backofficeTheme } from "../../../../../src/themes/backoffice";
-import { example } from "./affixes-example";
+import { ButtonNext } from "../../../../src/components/button-next";
+import { backofficeTheme } from "../../../../src/themes/backoffice";
+
+const example = `import * as React from "react";
+import { ButtonNext } from "wix-ui-core";
+import Sound from "wix-ui-icons-common/Sound";
+
+export default () => (
+  <React.Fragment>
+    <ButtonNext prefixIcon={<Sound />}>prefix</ButtonNext>
+    <ButtonNext suffixIcon={<Sound />}>suffix</ButtonNext>
+  </React.Fragment>
+);`;
 
 const description = (
   <div>
@@ -27,9 +37,6 @@ const Affixes = ({ style }: AffixesProps) => (
   >
     <ButtonNext prefixIcon={<Sound />}>prefix</ButtonNext>
     <ButtonNext suffixIcon={<Sound />}>suffix</ButtonNext>
-    <ButtonNext prefixIcon={<Sound />} suffixIcon={<Sound />}>
-      both
-    </ButtonNext>
   </CodeShowcase>
 );
