+++ packages/wix-ui-core/stories/backoffice/button-next/showcase/size.tsx
@@ -1,11 +1,20 @@
 import * as React from "react";
 import CodeShowcase from "wix-storybook-utils/CodeShowcase";
-import { ButtonNext } from "../../../../../src/components/button-next";
-import {
-  buttonNext,
-  backofficeTheme
-} from "../../../../../src/themes/backoffice";
-import { example } from "./size-example";
+import { ButtonNext } from "../../../../src/components/button-next";
+import { buttonNext, backofficeTheme } from "../../../../src/themes/backoffice";
+
+const example = `import * as React from "react";
+import { ButtonNext } from "wix-ui-core";
+import { buttonNext } from "wix-ui-core/themes/backoffice";
+
+export default () => (
+  <React.Fragment>
+    <ButtonNext className={buttonNext('tiny')}>tiny</ButtonNext>
+    <ButtonNext className={buttonNext('small')}>small</ButtonNext>
+    <ButtonNext className={buttonNext('medium')}>medium</ButtonNext>
+    <ButtonNext className={buttonNext('large')}>large</ButtonNext>
+  </React.Fragment>
+);`;
 
 const description = (
   <div>
