--- packages/wix-ui-core/stories/backoffice/button-next/showcase/size/size-example.tsx
@@ -1,12 +1,0 @@
-export const example = `import * as React from "react";
-import { ButtonNext } from "wix-ui-core";
-import { buttonNext } from "wix-ui-core/themes/backoffice";
-
-export default () => (
-  <React.Fragment>
-    <ButtonNext className={buttonNext('tiny')}>tiny</ButtonNext>
-    <ButtonNext className={buttonNext('small')}>small</ButtonNext>
-    <ButtonNext className={buttonNext('medium')}>medium</ButtonNext>
-    <ButtonNext className={buttonNext('large')}>large</ButtonNext>
-  </React.Fragment>
-);`;
