+++ packages/wix-ui-core/stories/backoffice/button-next/showcase/skin-secondary.tsx
@@ -1,17 +1,34 @@
 import * as React from "react";
 
 import CodeShowcase from "wix-storybook-utils/CodeShowcase";
-import { ButtonNext } from "../../../../../src/components/button-next";
-import {
-  buttonNext,
-  backofficeTheme
-} from "../../../../../src/themes/backoffice";
-import { example } from "./skin-secondary-example";
+import { ButtonNext } from "../../../../src/components/button-next";
+import { buttonNext, backofficeTheme } from "../../../../src/themes/backoffice";
+
+export const example = `import * as React from "react";
+import { ButtonNext } from "wix-ui-core";
+import { buttonNext } from "wix-ui-core/themes/backoffice";
+
+const premiumSecondary = buttonNext('premium', 'secondary');
+const darkSecondary = buttonNext('dark', 'secondary');
+const lightSecondary = buttonNext('light', 'secondary');
+const transparentSecondary = buttonNext('transparent, 'secondary');
+const destructiveSecondary = buttonNext('destructive', 'secondary');
+
+export default () => (
+  <React.Fragment>
+    <ButtonNext className={buttonNext('secondary')}>default</ButtonNext>
+    <ButtonNext className={destructiveSecondary}>destructive</ButtonNext>
+    <ButtonNext className={premiumSecondary}>premium</ButtonNext>
+    <ButtonNext className={darkSecondary}>dark</ButtonNext>
+    <ButtonNext className={lightSecondary}>light</ButtonNext>
+    <ButtonNext className={transparentSecondary}>transparent</ButtonNext>
+  </React.Fragment>
+);`;
 
 const description = (
   <div>
     Secondary skins <code>default</code>,<code>destructive</code>,
-    <code>premium</code>,<code>dark</code>, <code>light</code>,{" "}
+    <code>premium</code>,<code>dark</code>, <code>light</code>,
     <code>transparent</code>.
   </div>
 );
